requirements.txt
fastapi==0.123.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv
sqlalchemy
alembic
bcrypt
python-jose